Rails4 Support

+++ lib/schema_plus/active_record/connection_adapters/schema_statements.rb
@@ -4,7 +4,7 @@ module SchemaPlus::ActiveRecord::ConnectionAdapters
     def self.included(base) #:nodoc:
       base.class_eval do
         alias_method_chain :create_table, :schema_plus
-        alias_method_chain :add_index, :schema_plus
+        include AddIndex
       end
     end
 
@@ -25,36 +25,24 @@ module SchemaPlus::ActiveRecord::ConnectionAdapters
       # override rails' :force to cascade
       drop_table(table, if_exists: true, cascade: true) if options.delete(:force)
 
-      indexes = []
+      if ::ActiveRecord::VERSION::MAJOR.to_i < 4
+        indexes = []
+      end
       create_table_without_schema_plus(table, options) do |table_definition|
         table_definition.schema_plus_config = SchemaPlus.config.merge(config_options)
-        table_definition.name = table
+        if ::ActiveRecord::VERSION::MAJOR.to_i < 4
+          table_definition.name = table
+        end
         yield table_definition if block_given?
-        indexes = table_definition.indexes
+        if ::ActiveRecord::VERSION::MAJOR.to_i < 4
+          indexes = table_definition.indexes
+        end
       end
-      indexes.each do |index|
-        add_index(table, index.columns, index.opts)
+      if ::ActiveRecord::VERSION::MAJOR.to_i < 4
+        indexes.each do |index|
+          add_index(table, index.columns, index.opts)
+        end
       end
-    end
-
-    ##
-    # :method: add_index
-    #
-    # SchemaPlus modifies SchemaStatements::add_index so that it ignores
-    # errors raised about add an index that already exists -- i.e. that has
-    # the same index name, same columns, and same options -- and writes a
-    # warning to the log. Some combinations of rails & DB adapter versions
-    # would log such a warning, others would raise an error; with
-    # SchemaPlus all versions log the warning and do not raise the error.
-    #
-    # (This avoids collisions between SchemaPlus's auto index behavior and
-    # legacy explicit add_index statements, for platforms that would raise
-    # an error.)
-    #
-    def add_index_with_schema_plus(table, columns, options={})
-      add_index_without_schema_plus(table, columns, options)
-    rescue => e
-      SchemaStatements.add_index_exception_handler(self, table, columns, options, e)
     end
 
     def self.add_index_exception_handler(connection, table, columns, options, e) #:nodoc:
@@ -66,5 +54,32 @@ module SchemaPlus::ActiveRecord::ConnectionAdapters
       ::ActiveRecord::Base.logger.warn "[schema_plus] Index name #{name.inspect}' on table #{table.inspect} already exists. Skipping."
     end
 
+    module AddIndex
+      def self.included(base) #:nodoc:
+        base.class_eval do
+          alias_method_chain :add_index, :schema_plus
+        end
+      end
+
+      ##
+      # :method: add_index
+      #
+      # SchemaPlus modifies SchemaStatements::add_index so that it ignores
+      # errors raised about add an index that already exists -- i.e. that has
+      # the same index name, same columns, and same options -- and writes a
+      # warning to the log. Some combinations of rails & DB adapter versions
+      # would log such a warning, others would raise an error; with
+      # SchemaPlus all versions log the warning and do not raise the error.
+      #
+      # (This avoids collisions between SchemaPlus's auto index behavior and
+      # legacy explicit add_index statements, for platforms that would raise
+      # an error.)
+      #
+      def add_index_with_schema_plus(table, columns, options={})
+        add_index_without_schema_plus(table, columns, options)
+      rescue => e
+        SchemaStatements.add_index_exception_handler(self, table, columns, options, e)
+      end
+    end
   end
 end
